pyproject.toml
[project]
name = "django-project-structure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-redis>=5.4.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "markdown>=3.8",
    "psycopg2-binary>=2.9.10",
    "redis>=6.1.0",
]
